move ethereum error class and helpers to react-helpers
