+++ libs/react-helpers/src/index.ts
@@ -4,5 +4,6 @@ export * from './lib/grid-cells';
 export * from './lib/storage';
 export * from './lib/generic-data-provider';
 export * from './lib/i18n';
+export * from './lib/ethereum-error';
 
 export * from './hooks';
